keywordio/main/requirements.txt
anyio==3.5.0
appdirs==1.4.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.1
asttokens==2.0.5
attrs==21.4.0
awscli==1.22.55
AWSIoTPythonSDK==1.4.9
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
bleach==4.1.0
bootstrap4==0.1.0
boto3==1.20.53
botocore==1.24.0
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.10
click==8.0.3
Collectfast==2.2.0
colorama==0.4.4
ConnectWifiWithPy==0.0.1.0
contextlib2==0.6.0
debugpy==1.5.1
defusedxml==0.7.1
distlib==0.3.0
Django==3.1.4
django-admin-autocomplete-filter==0.7.1
django-environ==0.4.5
django-filter==21.1
django-jasmine==0.5.1
django-jazzmin==2.4.9
django-rest-passwordreset==1.2.1
django-simple-history==3.0.0
django-storages==1.12.3
django-truncate==0.1
djangorestframework==3.13.1
docutils==0.15.2
durationpy==0.5
environ==1.0
executing==0.8.2
gTTS==2.2.3
h11==0.12.0
hjson==3.0.2
html5lib==1.0.1
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipaddr==2.2.0
ipykernel==6.9.1
ipython==8.0.1
ipywidgets==7.6.5
jedi==0.18.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kappa==0.6.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==0.6.2
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.2
packaging==20.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pep517==0.8.2
pickleshare==0.7.5
Pillow==8.0.1
pip-tools==6.5.1
placebo==0.10.0
progress==1.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==2.4.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytoml==0.1.21
pytz==2020.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
retrying==1.3.3
rfc3986==1.5.0
s3transfer==0.5.1
scikit-learn==0.24.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
stack-data==0.2.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
troposphere==3.2.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.2
wincertstore==0.2
wsgi-request-logger==0.4.6
zappa==0.54.1
zipp==3.7.0
